Add support for Arduino Due's software erase hack
The Arduino Due's USB-to-Serial interface chip has special code that
will erase and reset the SAM3X8E if you open a 1200 baud connection and
immediately close it. This very simple change adds a command line option
to bossac to do that (the option is '-a'). If you use this option you
must specify the port.

I tested it on Windows and it works fine.