Short delay to ensure port is flushed in OS X (and prevent download hang)
on OS X, it appears that occasionally the _port->flush() may not always complete before writeBinary() is called.  This results in an exception and an error during download (as it looks like we're running into the "SAM firmware bug" mentioned in the comments.  Adding this short delay helps ensure that the flush() command competes before the binary write is called.