from: bugs.debian.org/953068
src/Command.cpp
In Ubuntu, the latest version of bossa has failed to build on ppc64el
because the ppc64el architecture in Ubuntu uses -O3 as a default
optimization, which exposes a number of uninitialized variables in the code.

In reality none of these variables are used uninitialized, but the compiler
can't /prove/ that, so it errors out.